virt-launcher monitor process
virt-launcher fork itself for monitoring purposes. This causes a `double` memory consumption that can be reduced.
Instead of fork itself, create a virt-launcher-monitor process that acts as an entrypoint and start the real virt-launcher process passing to it all the arguments.
The main purpose is to create a process that minimize the import of external libraries to minimize the impact on the memory usage.
