+++ cmd/virt-launcher/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//pkg/hooks:go_default_library",
         "//pkg/hotplug-disk:go_default_library",
         "//pkg/ignition:go_default_library",
-        "//pkg/network/istio:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-launcher:go_default_library",
@@ -28,12 +27,9 @@ go_library(
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/net:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
-        "//vendor/k8s.io/client-go/util/retry:go_default_library",
         "//vendor/libvirt.org/go/libvirt:go_default_library",
     ],
 )
@@ -62,6 +58,7 @@ pkg_tar(
         ":virt-launcher",
         "//cmd/container-disk-v2alpha:container-disk",
         "//cmd/virt-freezer",
+        "//cmd/virt-launcher-monitor",
         "//cmd/virt-probe",
     ],
     package_dir = "/usr/bin",
